refactor(report): split to three services
